:art: jdk代理不能有抽象类
